requirements.txt
Flask_Babel==2.0.0
Flask==1.1.2
Flask_Admin==1.5.8
flask_restplus==0.13.0
Flask_Testing==0.8.1
Flask_SQLAlchemy==2.5.1
Flask_Mail==0.9.1
Flask_Cors==3.0.10
boto3==1.17.69
itsdangerous==1.1.0
SQLAlchemy==1.4.11
Flask_Bcrypt==0.7.1
Flask_Script==2.0.6
requests==2.25.1
Flask_Migrate==2.7.0
alembic==1.5.8
langdetect==1.0.8
botocore==1.20.69
PyJWT==2.1.0
Werkzeug==0.16.1
psycopg2==2.8.6
